26.05.27 count5

+++ 2026_1/ML_Theory-Practice/course-ml-basic/260527/test.py
@@ -1,6 +1,6 @@
 import numpy as np
 
 # 2 x 1 x 3
-bar = np.array([2.0, 3.0], dtype=np.float32)
+bar = np.array([2.0, 3.0])
 
 print(bar.dtype)
